requirements.txt
asyncpg~=0.25.0
alembic~=1.8.0
ConfigArgParse~=1.5.3
fastapi~=0.78.0
psycopg2-binary==2.9.3
python-dateutil~=2.8.2
python-decouple~=3.6
SQLAlchemy==1.4.37
sqlmodel==0.0.6
setuptools~=60.2.0
uvicorn~=0.17.6
yarl~=1.7.2
pydantic~=1.9.1